fix warnings

+++ Engine/AppManager.cpp
@@ -418,7 +418,7 @@ AppManager::loadFromArgs(const CLArgs& cl)
     initializeQApp(_imp->nArgs, &_imp->commandLineArgsUtf8.front());
     // see C++ standard 23.2.4.2 vector capacity [lib.vector.capacity]
     // resizing to a smaller size doesn't free/move memory, so the data pointer remains valid
-    assert(_imp->nArgs <= _imp->commandLineArgsUtf8.size());
+    assert(_imp->nArgs <= (int)_imp->commandLineArgsUtf8.size());
     _imp->commandLineArgsUtf8.resize(_imp->nArgs); // Qt may have reduced the numlber of args
 
 #ifdef QT_CUSTOM_THREADPOOL
